requirements.txt
requests
beautifulsoup4
python-dotenv
Flask
curl_cffi
